server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
psycopg2-binary
Flask-RESTful
python-dotenv
gunicorn
openai
langchain-community
langgraph
langchain-anthropic
tavily-python
langgraph-checkpoint-sqlite